Update Padawan360_mega_maestro_DYSV5W_PWM.ino
Error found in line 81 definition of TURNSPEED by Jeff Cowman.
Variable should be declared as "float" rather than "byte".  This error meant that the turn calculation would not work.

+++ Padawan360_mega_maestro_DYSV5W_PWM.ino
@@ -78,7 +78,7 @@ byte drivespeed = DRIVESPEED1;
 // the higher this number the faster the droid will spin in place, lower - easier to control.
 // Recommend beginner: 40 to 50, experienced: 50 $ up, I like 70
 // This may vary based on your drive system and power system
-const byte TURNSPEED = 40;
+const float TURNSPEED = 40;
 
 // Set isLeftStickDrive to true for driving  with the left stick
 // Set isLeftStickDrive to false for driving with the right stick (legacy and original configuration)
